Fix unusual build/require errors
For some reason, Crystal suddenly started complaining that SWAny was undefined inside Sherwood. Simply adding "include Types" to the interior of Sherwood also gave access to other code as "Sherwood::SWAny" instead of "Sherwood::Types::SWAny". This is a bit of a hacky workaround, but fixes the compile errors without introducing added public constants to the Sherwood class.